requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Django==5.1.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
Dash==2.14.2
Dash-Bootstrap-Components==1.5.0
plotly==5.18.0
pandas==2.1.3
python-dateutil==2.8.2
gunicorn==21.2.0
